SASS-7584: Upgrade income-tax-additional-info-frontend to play 3.0
