+++ build.sbt
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import uk.gov.hmrc.DefaultBuildSettings.integrationTestSettings
+import scoverage.ScoverageKeys
+import uk.gov.hmrc.DefaultBuildSettings
 
 val appName = "income-tax-additional-info-frontend"
 
@@ -62,39 +62,47 @@ lazy val microservice = Project(appName, file("."))
   .enablePlugins(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin)
   .disablePlugins(sbt.plugins.JUnitXmlReportPlugin)
   .disablePlugins(JUnitXmlReportPlugin)
-  .settings(
-    // To resolve a bug with version 2.x.x of the scoverage plugin - https://github.com/sbt/sbt/issues/6997
-    libraryDependencySchemes ++= Seq("org.scala-lang.modules" %% "scala-xml" % VersionScheme.Always,
-    )
-  )
   .settings(PlayKeys.playDefaultPort := 10005)
   .settings(
     majorVersion := 0,
     scalaVersion := "2.13.12",
-    libraryDependencies ++= AppDependencies.compile ++ AppDependencies.test,
+    libraryDependencies ++= AppDependencies.compile ++ AppDependencies.test ++ AppDependencies.itDependencies,
     TwirlKeys.templateImports ++= twirlImports,
-    Assets / pipelineStages := Seq(gzip),
+    // only required for frontends
     scalacOptions += "-Wconf:cat=unused-imports&src=html/.*:s",
-    scalacOptions += "-Wconf:src=routes/.*:s"
+    // for all services
+    scalacOptions += "-Wconf:cat=unused-imports&src=routes/.*:s"
+   //scalacOptions += "-Wconf:src=routes/.*:s"
   )
   .settings(Test / fork := false)
-  .configs(IntegrationTest extend Test)
-  .settings(integrationTestSettings(): _*)
+  .settings(
+    DefaultBuildSettings.itSettings() ++
+      Seq(unmanagedResourceDirectories.withRank(KeyRanks.Invisible) := Seq(baseDirectory.value / "it" / "resources")
+      ))
   .settings(resolvers += Resolver.jcenterRepo)
-  .settings(IntegrationTest / unmanagedResourceDirectories += baseDirectory.value / "it" / "resources")
   .settings(coverageSettings: _*)
   .settings(
     // concatenate js
-    Concat.groups := Seq("javascripts/application.js" -> group(Seq(
-      "lib/govuk-frontend/govuk/all.js",
-      "javascripts/jquery.min.js",
-      "javascripts/app.js",
-    ))),
-    // prevent removal of unused code which generates warning errors due to use of third-party libs
-    uglifyCompressOptions := Seq("unused=false", "dead_code=false"),
+    Concat.groups := Seq(
+      "javascripts/application.js" ->
+        group(Seq(
+            "lib/govuk-frontend/govuk/all.js",
+            "javascripts/jquery.min.js",
+            "javascripts/app.js"
+          )
+        )
+    ),
     pipelineStages := Seq(digest),
     // below line required to force asset pipeline to operate in dev rather than only prod
-    Assets / pipelineStages := Seq(concat, uglify),
-    // only compress files generated by concat
-    uglify / includeFilter := GlobFilter("application.js")
+    Assets / pipelineStages := Seq(concat)
   )
+
+lazy val it = project
+  .enablePlugins(PlayScala)
+  .dependsOn(microservice % "test->test") // the "test->test" allows reusing test code and test dependencies
+  .settings(DefaultBuildSettings.itSettings())
+  .settings(scalaVersion := "2.13.12")
+  .settings(majorVersion := 0)
+  .settings(libraryDependencies ++= AppDependencies.test)
+
+addCommandAlias("runAllChecks", "clean;compile;scalastyle;coverage;test;it/test;coverageReport")
