+++ test/support/providers/ResultBodyConsumerProvider.scala
@@ -16,7 +16,7 @@
 
 package support.providers
 
-import akka.actor.ActorSystem
+import org.apache.pekko.actor.ActorSystem
 import play.api.mvc.Result
 import play.api.test.DefaultAwaitTimeout
 import play.api.test.Helpers.await
